+++ lib/go-atscfg/volumedotconfig.go
@@ -23,6 +23,8 @@ import (
 	"strconv"
 )
 
+const ContentTypeVolumeDotConfig = ContentTypeTextASCII
+
 // MakeVolumeDotConfig creates volume.config for a given ATS Profile.
 // The paramData is the map of parameter names to values, for all parameters assigned to the given profile, with the config_file "storage.config".
 func MakeVolumeDotConfig(