+++ lib/go-tc/deliveryservices.go
@@ -165,8 +165,10 @@ type DeliveryServiceNullableV15 DeliveryServiceNullable // this type alias shoul
 
 type DeliveryServiceNullable struct {
 	DeliveryServiceNullableV14
-	EcsEnabled bool `json:"ecsEnabled" db:"ecs_enabled"`
+	EcsEnabled          bool `json:"ecsEnabled" db:"ecs_enabled"`
+	RangeSliceBlockSize *int `json:"rangeSliceBlockSize" db:"range_slice_block_size"`
 }
+
 type DeliveryServiceNullableV14 struct {
 	DeliveryServiceNullableV13
 	ConsistentHashRegex       *string  `json:"consistentHashRegex"`
@@ -392,6 +394,23 @@ func (ds *DeliveryServiceNullable) validateTypeFields(tx *sql.Tx) error {
 		"orgServerFqdn": validation.Validate(ds.OrgServerFQDN,
 			validation.By(requiredIfMatchesTypeName([]string{DNSRegexType, HTTPRegexType}, typeName)),
 			validation.NewStringRule(validateOrgServerFQDN, "must start with http:// or https:// and be followed by a valid hostname with an optional port (no trailing slash)")),
+		"rangeSliceBlockSize": validation.Validate(ds,
+			validation.By(func(dsi interface{}) error {
+				ds := dsi.(*DeliveryServiceNullable)
+				if ds.RangeRequestHandling != nil {
+					if *ds.RangeRequestHandling == 3 {
+						return validation.Validate(ds.RangeSliceBlockSize, validation.Required,
+							// Per Slice Plugin implementation
+							validation.Min(262144),   // 256KiB
+							validation.Max(33554432), // 32MiB
+						)
+					}
+					if ds.RangeSliceBlockSize != nil {
+						return errors.New("rangeSliceBlockSize can only be set if the rangeRequestHandling is set to 3 (Use the Slice Plugin)")
+					}
+				}
+				return nil
+			})),
 		"protocol": validation.Validate(ds.Protocol,
 			validation.By(requiredIfMatchesTypeName([]string{SteeringRegexType, DNSRegexType, HTTPRegexType}, typeName))),
 		"qstringIgnore": validation.Validate(ds.QStringIgnore,