+++ traffic_ops/traffic_ops_golang/api/api.go
@@ -94,7 +94,7 @@ func WriteRespRaw(w http.ResponseWriter, r *http.Request, v interface{}) {
 		tc.GetHandleErrorsFunc(w, r)(http.StatusInternalServerError, errors.New(http.StatusText(http.StatusInternalServerError)))
 		return
 	}
-	w.Header().Set("Content-Type", "application/json")
+	w.Header().Set(rfc.ContentType, rfc.ApplicationJSON)
 	w.Write(append(bts, '\n'))
 }
 
@@ -136,6 +136,14 @@ func HandleErr(w http.ResponseWriter, r *http.Request, tx *sql.Tx, statusCode in
 		}
 	}
 	handleSimpleErr(w, r, statusCode, userErr, sysErr)
+}
+
+func HandleErrOptionalDeprecation(w http.ResponseWriter, r *http.Request, tx *sql.Tx, statusCode int, userErr error, sysErr error, deprecated bool, alternative *string) {
+	if deprecated {
+		HandleDeprecatedErr(w, r, tx, statusCode, userErr, sysErr, alternative)
+	} else {
+		HandleErr(w, r, tx, statusCode, userErr, sysErr)
+	}
 }
 
 // HandleDeprecatedErr handles an API error, adding a deprecation alert, rolling back the transaction, writing the given statusCode and userErr to the user, and logging the sysErr. If userErr is nil, the text of the HTTP statusCode is written.
@@ -279,7 +287,7 @@ func WriteRespAlertObj(w http.ResponseWriter, r *http.Request, level tc.AlertLev
 		return
 	}
 	w.Header().Set(rfc.ContentType, rfc.ApplicationJSON)
-	w.Write(append(respBts, '\n'))
+	_, _ = w.Write(append(respBts, '\n'))
 }
 
 func WriteAlerts(w http.ResponseWriter, r *http.Request, code int, alerts tc.Alerts) {
@@ -289,17 +297,24 @@ func WriteAlerts(w http.ResponseWriter, r *http.Request, code int, alerts tc.Ale
 	}
 	setRespWritten(r)
 
-	resp, err := json.Marshal(alerts)
-	if err != nil {
-		handleSimpleErr(w, r, http.StatusInternalServerError, nil, fmt.Errorf("marshalling JSON: %v", err))
-		return
-	}
 	w.Header().Set(rfc.ContentType, rfc.ApplicationJSON)
 	w.WriteHeader(code)
-	w.Write(append(resp, '\n'))
+	if alerts.HasAlerts() {
+		resp, err := json.Marshal(alerts)
+		if err != nil {
+			handleSimpleErr(w, r, http.StatusInternalServerError, nil, fmt.Errorf("marshalling JSON: %v", err))
+			return
+		}
+		_, _ = w.Write(append(resp, '\n'))
+	}
 }
 
 func WriteAlertsObj(w http.ResponseWriter, r *http.Request, code int, alerts tc.Alerts, obj interface{}) {
+	if !alerts.HasAlerts() {
+		WriteResp(w, r, obj)
+		w.WriteHeader(code)
+		return
+	}
 	if respWritten(r) {
 		log.Errorf("WriteAlertsObj called after a write already occurred! Not double-writing! Path %s", r.URL.Path)
 		return